restructure files
restructure /test
