+++ test/calibration.jl
@@ -5,82 +5,6 @@ using IterationControl
 import Lux
 using Plots
 
-times = [0, 10, 30]
-volumes = [3, 6, 15]
-
-@testset "guess_parameters" begin
-    # classical:
-    p = TumorGrowth.guess_parameters(times, volumes, gompertz)
-    @test length(p) == 3
-    @test p.v0 ≈ 3
-    @test p.v∞ == 15
-    @test p.ω ≈ (log(15) - log(3))/(30)
-
-    # one-dimensional TumorGrowth.
-    p = TumorGrowth.guess_parameters(times, volumes, bertalanffy)
-    @test length(p) == 4
-    @test p.v0 ≈ 3
-    @test p.v∞ == 15
-    @test p.ω ≈ (log(15) - log(3))/(30)
-    @test p.λ ≈ 1/3
-
-    # two-d TumorGrowt (just a smoke test).
-    p = TumorGrowth.guess_parameters(times, volumes, bertalanffy2)
-    @test length(p) == 5
-    keys(p) == (:v∞, :ω, :λ, :γ)
-
-    # fallback:
-    @test isnothing(TumorGrowth.guess_parameters(times, volumes, "junk"))
-end
-
-@testset "scale_function" begin
-    # classical:
-    s1 = TumorGrowth.scale_function(times, volumes, gompertz)
-    scales = s1((v0=1.0, v∞=1.0, ω=1.0))
-    @test scales.v0 == 15.0
-    @test scales.v∞ == 15.0
-    @test scales.ω ≈ abs.(TumorGrowth.guess_parameters(times, volumes, gompertz).ω)
-
-    # one-dimensional TumorGrowth.
-    s1 = TumorGrowth.scale_function(times, volumes, bertalanffy)
-    scales = s1((v0=1.0, v∞=1.0, ω=1.0, λ=1.0))
-    @test scales.v0 == 15.0
-    @test scales.v∞ == 15.0
-    @test scales.ω ≈ abs.(TumorGrowth.guess_parameters(times, volumes, bertalanffy).ω)
-    @test scales.λ == 1.0
-
-    # bertalanffy2
-    s1 = TumorGrowth.scale_function(times, volumes, bertalanffy2)
-    p = guess_parameters(times, volumes, bertalanffy2)
-    vol_scale = p.v∞
-    scales = s1((v0=1.0, v∞=1.0, ω=1.0, λ=1.0, γ=1.0))
-    @test scales.v0 ≈ vol_scale
-    @test scales.v∞ ≈ vol_scale
-    @test scales.ω ≈ abs.(TumorGrowth.guess_parameters(times, volumes, bertalanffy2).ω)
-    @test scales.λ == 1.0
-    @test scales.γ == 1.0
-
-    # fallback:
-    @test TumorGrowth.scale_function(times, volumes, "junk") == identity
-end
-
-@testset "constraint function" begin
-    for model in [gompertz, bertalanffy, bertalanffy2]
-        h = TumorGrowth.constraint_function(model)
-        p = TumorGrowth.guess_parameters(times, volumes, model)
-        @test h(p)
-        q = merge(p, (; v∞ = -p.v∞))
-        @test !h(q)
-        q = merge(p, (; v∞ = -p.v∞, v0 = -p.v0))
-        @test !h(q)
-        q = merge(p, (; v∞ = -p.v∞))
-        @test !h(q)
-    end
-
-    h = TumorGrowth.constraint_function("junk")
-    @test h("a;lsjfd")
-end
-
 function normalized_absolute_difference(yhat, y)
     scale = abs(y)
     abs(yhat - y)/scale
